Move DaikinNNN code into separate files.  No other fixes yet.
